Extract common module to separate file refactor
